requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pytesseract==0.3.10
Pillow==10.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
regex==2023.8.8
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
